+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-VERSION HISTORY 
+VERSION HISTORY
+
+VERSION 0.2.0 - 01/24/2026
+ - Time trial ghost mechanics added
+ - Ghost that becomes player's best time in the race
 
 VERSION 0.1.0 - 01/21/2026
  - Basic structure of game created